requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.11
